Add Python Dependabot Configuration (#9)
# Pull Request

## Description

This pull request adds a new configuration for Dependabot in the
`.github/dependabot.yml` file to manage updates for the `uv` package
ecosystem. The changes introduce a scheduled cron job and grouping for
version updates.

### Dependabot configuration enhancements:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R22-R37):
Added a new `uv` package ecosystem configuration with a cron schedule
set to "30 7 * * *" in the "Europe/London" timezone. This configuration
targets the `main` branch and includes grouping for version updates with
patterns and update types specified as "patch" and "minor."